+++ src/card/cards.rs
@@ -4,9 +4,11 @@ use crate::CardDefinitionId;
 
 mod isd_rtr;
 mod landwalk;
+mod rampage;
 mod regeneration;
 pub use isd_rtr::*;
 pub use landwalk::*;
+pub use rampage::*;
 pub use regeneration::*;
 
 pub const ANKH_OF_MISHRA: CardDefinitionId = CardDefinitionId(1);